usb: tps6598x: Wait for registers to stablize after IRQ
After cable insert / register status interrupts, the TPS6598x
may still be in the middle of determining the type of cable that is
attached.  So if we use the register values immediately after an IRQ,
we may end up notifying the system of a cable state that was only
an intermediate step in detection.

Since there is no method apparent that lets SW know when detection
is actually complete, make a best attempt to allow the status
registers to stabilze before we nofity the system of the state change.