usb: tps6598x: Optionally notify USB power_supply of data roles
Since Type C allows for UFP or DFP mode for either power sinks or
power sources, need to communicate the data role of the system
to the USB PHY driver based on cable detection.

This will default the system to host mode unless a UFP cable with a USB
host on the far end is detected.