ASoC: tlv320adc5140: Add the tlv320adc5140 codec driver
Add the tlv320adc5140 codec driver.
This is a test driver and is not complete in the mapping or the controls.
There is an init hack which hard enables the 4 Analog microphones.

THIS DRIVER SHOULD NOT BE USED FOR ANY PRODUCTION.

There is also a debug file that allows the reading and writing of the registers.

Signed-off-by: Dan Murphy <dmurphy@ti.com>